delete unhandle min and max detections js
